Externalize inlines justification
