+++ src/components/core/InlineManager.js
@@ -15,6 +15,7 @@ in its own updateLayout function.
  */
 import * as Whitespace from '../../utils/inline-layout/Whitespace';
 import * as TextAlign from '../../utils/inline-layout/TextAlign';
+import * as InlineJustification from '../../utils/inline-layout/InlineJustification';
 
 export default function InlineManager( Base ) {
 
@@ -38,47 +39,8 @@ export default function InlineManager( Base ) {
 			// Position lines according to justifyContent and contentAlign
 			/////////////////////////////////////////////////////////////////
 
-			// individual vertical offset
-
-			const textHeight = Math.abs( lines.height );
-
-
-			// Line vertical positioning
-
-			const justificationOffset = ( () => {
-				switch ( JUSTIFICATION ) {
-
-					case 'start':
-						return ( INNER_HEIGHT / 2 ) - lines[ 0 ].lineHeight;
-
-					case 'end':
-						return textHeight - lines[ 0 ].lineHeight - ( INNER_HEIGHT / 2 ) + ( lines[ lines.length - 1 ].lineHeight - lines[ lines.length - 1 ].lineHeight );
-
-					case 'space-around':
-					case 'space-between':
-					case 'space-evenly':
-					case 'center':
-						return ( textHeight / 2 ) - lines[ 0 ].lineHeight;
-
-					default:
-						console.warn( `justifyContent: '${JUSTIFICATION}' is not valid` );
-
-				}
-			} )();
-
-
-			//
-
-			lines.forEach( ( line ) => {
-
-				line.y += justificationOffset;
-				line.forEach( ( inline ) => {
-
-					inline.offsetY += justificationOffset;
-
-				} );
-
-			} );
+			// Vertical positioning
+			InlineJustification.justifyInlines( lines, JUSTIFICATION, INNER_HEIGHT);
 
 			// Horizontal positioning
 			TextAlign.textAlign( lines, ALIGNMENT, INNER_WIDTH );
